+++ app/controllers/auth.py
@@ -39,13 +39,13 @@ def githubCallback():
     session['access_token'] = access_token
     session['user_id'] = user.id
 
-    return redirect(url_for('home.index'))
+    return redirect(url_for('art.index'))
 
 
 @blueprint.route('/logout')
 def logout():
     session.clear()
-    return redirect(url_for('home.index'))
+    return redirect(url_for('art.index'))
 
 
 @blueprint.before_app_request
