+++ app/templates/art/index.html
@@ -6,114 +6,65 @@
 {% endblock %}
 
 {% block content %}
-<div class="container p-0">
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
-  <div></div>
+<div class="header-form mt-4">
+  {% if g.user %}
+  <form action="/art/" method="POST" class="form-inline" enctype="multipart/form-data">
+    <div class="nav-item p-2">
+      <form id="form" action="/" method="post" enctype="multipart/form-data"
+        class="form-inline d-flex align-items-center justify-content-center">
+        <input type="file" name="file" id="file" class="form-control mr-2">
+        <button class="btn btn-light" type="submit">Upload</button>
+      </form>
+      <a class="btn btn-dark text-white" href="{{ url_for('auth.logout') }}">Log out</a>
+    </div>
+    {% else %}
+    <div class="nav-item">
+      <a class="btn btn-dark text-white" href="{{ url_for('auth.githubLogin') }}">
+        Login with GitHub
+      </a>
+    </div>
+  </form>
+  {% endif %}
+  {% for category, message in get_flashed_messages(with_categories=true) %}
+  <div class="alert alert-{{ category if category != None else 'danger' }} mb-0 rounded-0">
+    <div class="container">
+      {{ message }}
+    </div>
+  </div>
+  {% endfor %}
 </div>
-<script src="https://code.jquery.com/jquery-3.4.1.js"></script>
-<script type="text/javascript">
-  $(document).ready(function () {
-    $('.container div').mouseover(function () {
-      $(this).addClass('flip')
-    })
-  })
-</script>
+<div id="indicators" class="carousel slide" data-ride="carousel">
+  <ol class="carousel-indicators">
+    <li data-target="#indicators" data-slide-to="0" class="active"></li>
+    <li data-target="#indicators" data-slide-to="1"></li>
+    {% for photo in photos %}
+    <li data-target="#indicators" data-slide-to="{{photo.id}}"></li>
+    {% endfor %}
+  </ol>
+  <div class="carousel-inner">
+    <div class="carousel-item active">
+      <img class="d-block w-100"
+        src="https://images.unsplash.com/photo-1617228133035-2347f159e755?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=1350&q=80"
+        alt="First slide">
+    </div>
+    <div class="carousel-item">
+      <img class="d-block w-100"
+        src="https://images.unsplash.com/photo-1617252820859-00a22c77ec0c?ixid=MXwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHw%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=634&q=80"
+        alt="Second slide">
+    </div>
+    {% for photo in photos %}
+    <div class="carousel-item">
+      <img class="d-block w-100" src="..{{photo.image_url}}" alt="{{photo.id}}">
+    </div>
+    {% endfor %}
+  </div>
+  <a class="carousel-control-prev" href="#indicators" role="button" data-slide="prev">
+    <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+    <span class="sr-only">Previous</span>
+  </a>
+  <a class="carousel-control-next" href="#indicators" role="button" data-slide="next">
+    <span class="carousel-control-next-icon" aria-hidden="true"></span>
+    <span class="sr-only">Next</span>
+  </a>
+</div>
 {% endblock %}
