+++ app/templates/layouts/layout.html
@@ -15,8 +15,6 @@
   <body>
     {% include "partials/header.html" %}
 
-    {% include "partials/flash.html" %}
-
     {% block header %}{% endblock %}
 
     {% block content %}{% endblock %}
@@ -38,12 +36,14 @@
         //jquery for expand and collapse the sidebar
         $('.menu-btn').click(function () {
           $('.side-bar').addClass('active');
+          $('.side-bar').css("z-index", "10000");
           $('.menu-btn').css("visibility", "hidden");
         });
 
         $('.close-btn').click(function () {
           $('.side-bar').removeClass('active');
           $('.menu-btn').css("visibility", "visible");
+          $('.side-bar').css("z-index", "0");
         });
       });
     </script>
